refactor: image custom 분리
